use Blueprint on flask, Add loading processing
