+++ static/homepage.css
@@ -318,6 +318,15 @@
   object-fit: cover;
 }
 
+.loading-img {
+  height: 75%;
+  width: 100%;
+  background-image: url("/static/images/Loading-img.svg");
+  background-size: 30%; 
+  background-repeat: no-repeat; 
+  background-position: center center; 
+}
+
 .attraction-name-box {
   position: absolute;
   bottom: 0;
@@ -338,7 +347,6 @@
 
 .name-text {
   color: var(--additional);
-
 }
 
 .attraction-text-container {
@@ -373,7 +381,6 @@
   bottom: 0;
 }
 
-
 /* not found box-----------------------------------*/
 .not-found-box {
   color: var(--grayColor);
@@ -396,7 +403,6 @@
 .footer-text {
   color: var(--additional);
 }
-
 
 @media screen and (max-width: 1199px) {
   .nav-container {
@@ -433,7 +439,6 @@
     grid-template-columns: repeat(2, 1fr);
     gap: 15px;
     width: 95%;
-
   }
 
   .attraction-container {
@@ -443,7 +448,6 @@
 }
 
 @media screen and (max-width: 599px) {
-
   .main-container-attractions {
     display: flex;
     flex-direction: column;
